Move `VerificationStatus` to storage
