--- solarkraft/src/VerificationStatus.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * [Apache-2.0](https://github.com/freespek/solarkraft/blob/main/LICENSE)
- */
-export enum VerificationStatus {
-    NoViolation = 0,
-    Violation = 1,
-    Unknown = 2,
-}
